+++ lib/Table/TableStyle.ts
@@ -1,6 +1,6 @@
 import groupBy from "lodash-es/groupBy";
-import sortedIndex from "lodash-es/sortedIndex";
 import { computed } from "mobx";
+import binarySearch from "terriajs-cesium/Source/Core/binarySearch";
 import Color from "terriajs-cesium/Source/Core/Color";
 import JulianDate from "terriajs-cesium/Source/Core/JulianDate";
 import TimeInterval from "terriajs-cesium/Source/Core/TimeInterval";
@@ -28,7 +28,6 @@ import TableTraits from "../Traits/TableTraits";
 import ColorPalette from "./ColorPalette";
 import TableColumn from "./TableColumn";
 import TableColumnType from "./TableColumnType";
-import binarySearch from "terriajs-cesium/Source/Core/binarySearch";
 
 const getColorForId = createColorForIdTransformer();
 const defaultColor = "yellow";
@@ -250,7 +249,7 @@ export default class TableStyle {
     }
 
     let paletteName = this.colorTraits.colorPalette;
-    let numberOfBins: number | undefined;
+    const numberOfBins = this.numberOfBins;
 
     if (
       colorColumn.type === TableColumnType.enum ||
@@ -259,7 +258,6 @@ export default class TableStyle {
     ) {
       // Enumerated values, so use a large, high contrast palette.
       paletteName = paletteName || "HighContrast";
-      numberOfBins = colorColumn.uniqueValues.values.length;
     } else if (colorColumn.type === TableColumnType.scalar) {
       if (paletteName === undefined) {
         const valuesAsNumbers = colorColumn.valuesAsNumbers;
@@ -275,7 +273,6 @@ export default class TableStyle {
           paletteName = "YlOrRd";
         }
       }
-      numberOfBins = this.binMaximums.length;
     }
 
     if (paletteName !== undefined && numberOfBins !== undefined) {
@@ -285,13 +282,31 @@ export default class TableStyle {
     }
   }
 
+  @computed
+  get numberOfBins(): number {
+    const colorColumn = this.colorColumn;
+    if (colorColumn === undefined) return this.binMaximums.length;
+    if (
+      colorColumn.type === TableColumnType.enum ||
+      colorColumn.type === TableColumnType.region ||
+      colorColumn.type === TableColumnType.text
+    ) {
+      return colorColumn.uniqueValues.values.length;
+    } else if (colorColumn.type === TableColumnType.scalar) {
+      return colorColumn.uniqueValues.values.length < this.binMaximums.length
+        ? colorColumn.uniqueValues.values.length
+        : this.binMaximums.length;
+    }
+    return this.binMaximums.length;
+  }
+
   /**
    * Gets the color to use for each bin. The length of the returned array
    * will be equal to {@link #numberOfColorBins}.
    */
   @computed
   get binColors(): readonly Readonly<Color>[] {
-    const numberOfBins = this.binMaximums.length;
+    const numberOfBins = this.numberOfBins;
 
     // Pick a color for every bin.
     const binColors = this.colorTraits.binColors || [];
@@ -334,8 +349,10 @@ export default class TableStyle {
       if (min === undefined || max === undefined) {
         return [];
       }
-
-      const numberOfBins = this.colorTraits.numberOfBins;
+      const numberOfBins =
+        colorColumn.uniqueValues.values.length < this.colorTraits.numberOfBins
+          ? colorColumn.uniqueValues.values.length
+          : this.colorTraits.numberOfBins;
       let next = min;
       const step = (max - min) / numberOfBins;