+++ CHANGES.md
@@ -14,6 +14,8 @@ Change Log
 * Temporarily stopped using `papaparse` for fetching Csv urls till an upstream bug is fixed.
 * Fix async bug with loading `ReferenceMixin` and then `Mappable` items in `initSources`
 * Remove `addToWorkbench`, it has been replaced with `workbench.add`
+* Improve handling of `ArcGisMapServerCatalogItem` when dealing with tiled layers.
+* Ensure there aren't more bins than unique values for a `TableStyle`
 
 #### 8.0.0-alpha.60
 * Fix WMS legend for default styles.